Make order of collected files reproducible
The output of the Groovy depends on the order of source files, e.g. when
generating stubs for a class and its superclass, `implements
GroovyObject` will be added to the class only if the superclass hasn't
already been processed. Therefore, this commit orders the files
lexicographically to make the output reproducible.